+++ pallets/ah-migrator/src/lib.rs
@@ -319,6 +319,8 @@ pub mod pallet {
 
 		/// Calls that are allowed after the migration finished.
 		type AhPostMigrationCalls: Contains<<Self as frame_system::Config>::RuntimeCall>;
+
+		type RcStakingMessage: Parameter + Convert2<Self::RcStakingMessage, AhEquivalentStakingMessageOf<Self>>;
 	}
 
 	/// RC accounts that failed to migrate when were received on the Asset Hub.
